refactor(chat-sessions): extract session query operation.
